+++ src/main/java/com/incidents/services/impl/GraffitiRemovalService.java
@@ -23,8 +23,8 @@ public class GraffitiRemovalService {
 	GraffitiRemovalDAO dao;
 	
 	@Transactional
-	public GraffitiRemoval create(String status, String streetAddress, String zipCode, Integer xCoordinate,
-			Integer yCoordinate, Integer ward, Integer policeDistrict, Integer communityArea, Integer latitude, Integer longitude,
+	public GraffitiRemoval create(String status, String streetAddress, String zipCode, Double xCoordinate,
+			Double yCoordinate, Integer ward, Integer policeDistrict, Integer communityArea, Double latitude, Double longitude,
 			String location, String typeOfSurface, String located, String ssa) {
 		Incident newIncident = new Incident();
 		newIncident.setCreatedAt(new Date());
