+++ src/main/java/com/incidents/web/GarbageCartsEndpoint.java
@@ -28,8 +28,8 @@ public class GarbageCartsEndpoint {
 	  private GarbageCartsService service;
 	  
 	  @RequestMapping(value = "/garbageCartsEndpoint/create", method = RequestMethod.POST, headers = "Accept=application/json")
-	  public GarbageCarts create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Integer xCoordinate,
-			  @RequestParam("yCoordinate") Integer yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Integer latitude, @RequestParam("longitude") Integer longitude,
+	  public GarbageCarts create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Double xCoordinate,
+			  @RequestParam("yCoordinate") Double yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Double latitude, @RequestParam("longitude") Double longitude,
 			  @RequestParam("location") String location, @RequestParam("blackCartsDelivered") Integer blackCartsDelivered, @RequestParam("currentActivity") String currentActivity,
 			  @RequestParam("mostRecentAction") String mostRecentAction, @RequestParam("ssa") String ssa) {
 		  MyUser user;
