+++ src/main/java/com/incidents/web/SanitationCodeComplaintsEndpoint.java
@@ -28,8 +28,8 @@ public class SanitationCodeComplaintsEndpoint {
 	  private SanitationCodeComplaintsService service;
 	  
 	  @RequestMapping(value = "/sanitationCodeComplaintsEndpoint/create", method = RequestMethod.POST, headers = "Accept=application/json")
-	  public SanitationCodeComplaints create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Integer xCoordinate,
-			  @RequestParam("yCoordinate") Integer yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Integer latitude, @RequestParam("longitude") Integer longitude,
+	  public SanitationCodeComplaints create(Principal principal, @RequestParam("status") String status, @RequestParam("streetAddress") String streetAddress, @RequestParam("zipCode") String zipCode, @RequestParam("xCoordinate") Double xCoordinate,
+			  @RequestParam("yCoordinate") Double yCoordinate, @RequestParam("ward") Integer ward, @RequestParam("policeDistrict") Integer policeDistrict, @RequestParam("communityArea") Integer communityArea, @RequestParam("latitude") Double latitude, @RequestParam("longitude") Double longitude,
 			  @RequestParam("location") String location, @RequestParam("natureOfViolation") String natureOfViolation) {
 		  MyUser user;
